ISSUE-294 Links refactor
A link can have an embedded component and a list component,
this is an attempt to stop some dupe in the database.

A link has data related to being embedded in the article,
which may be null (list_identifier='')

A link has data related to being in the links list,
which may be null (priority=0).

A link could have neither (makes no sense), or both (makes a lot
of sense, we can now have one record for both the embedded
official video and the link list equivalent)